source/requirements.txt
Django~=3.2
django-orderable-inlines
Pillow